common/requirements.txt
flask>=3.0
gunicorn>=21.0
PyJWT>=2.8
boto3>=1.34
